Fix logical connective `&&`
Currently in all L₂—L₈, both `&&` and `||` parse to `BOr`.
I believe this is unintentional as there is a `BAnd` defined but
not used.
It also causes confusion (it cost me a while of head-scratching
trying to understand why some of my experiments weren't typechecking
before I realized this was going on).